+++ captcha/tests/tests.py
@@ -1,54 +1,45 @@
-import os
-import warnings
-
 try:
-    from unittest.mock import patch
+    from importlib import reload
 except ImportError:
-    from mock import patch
+    pass
 
-from captcha import fields
-from django.forms import Form
+from django.conf import settings
+from django.core.exceptions import ImproperlyConfigured
 from django.test import TestCase, override_settings
 
-from captcha.client import RecaptchaResponse
+import captcha
 
 
-class TestForm(Form):
-    captcha = fields.ReCaptchaField(attrs={'theme': 'white'})
+class TestInit(TestCase):
 
-
-class TestCase(TestCase):
-    # No longer supports reCAPTCHA v1, override settings during tests to always
-    # use v2 reCAPTCHA. Prevents HTTP 410 error.
-    @override_settings(NOCAPTCHA=True)
-    @patch("captcha.fields.client.submit")
-    def test_client_success_response(self, mocked_submit):
-        mocked_submit.return_value = RecaptchaResponse(is_valid=True)
-        form_params = {'recaptcha_response_field': 'PASSED'}
-        form = TestForm(form_params)
-        self.assertTrue(form.is_valid())
-
-    # No longer supports reCAPTCHA v1, override settings during tests to always
-    # use v2 reCAPTCHA. Prevents HTTP 410 error.
-    @override_settings(NOCAPTCHA=True)
-    @patch("captcha.fields.client.submit")
-    def test_client_failure_response(self, mocked_submit):
-        mocked_submit.return_value = RecaptchaResponse(is_valid=False, error_code="410")
-        form_params = {'recaptcha_response_field': 'PASSED'}
-        form = TestForm(form_params)
-        self.assertFalse(form.is_valid())
-
-    # No longer supports reCAPTCHA v1, override settings during tests to always
-    # use v2 reCAPTCHA. Prevents HTTP 410 error.
-    @override_settings(NOCAPTCHA=True)
-    def test_deprecation_warning(self):
-        with warnings.catch_warnings(record=True) as w:
-            warnings.simplefilter("always")
-            form_params = {'recaptcha_response_field': 'PASSED'}
-            form = TestForm(form_params)
-
-            # Trigger warning on client.submit
-            form.is_valid()
-            assert len(w) == 1
-            assert issubclass(w[-1].category, DeprecationWarning)
-            assert "reCAPTCHA v1 will no longer be" in str(w[-1].message)
+    def test_setting_instance_check(self):
+        with override_settings(RECAPTCHA_PROXY="not a dict"):
+            with self.assertRaises(ImproperlyConfigured) as error:
+                reload(captcha)
+            self.assertEqual(error.exception.args, (
+                "Setting RECAPTCHA_PROXY is not of type", dict)
+            )
+        with override_settings(RECAPTCHA_VERIFY_REQUEST_TIMEOUT="not an int"):
+            with self.assertRaises(ImproperlyConfigured) as error:
+                reload(captcha)
+            self.assertEqual(error.exception.args, (
+                "Setting RECAPTCHA_VERIFY_REQUEST_TIMEOUT is not of type", int)
+            )
+        with override_settings(RECAPTCHA_DOMAIN=1):
+            with self.assertRaises(ImproperlyConfigured) as error:
+                reload(captcha)
+            self.assertEqual(error.exception.args, (
+                "Setting RECAPTCHA_DOMAIN is not of type", str)
+            )
+        with override_settings(RECAPTCHA_PUBLIC_KEY=1):
+            with self.assertRaises(ImproperlyConfigured) as error:
+                reload(captcha)
+            self.assertEqual(error.exception.args, (
+                "Setting RECAPTCHA_PUBLIC_KEY is not of type", str)
+            )
+        with override_settings(RECAPTCHA_PRIVATE_KEY=1):
+            with self.assertRaises(ImproperlyConfigured) as error:
+                reload(captcha)
+            self.assertEqual(error.exception.args, (
+                "Setting RECAPTCHA_PRIVATE_KEY is not of type", str)
+            )